+++ src/settings/Settings.ts
@@ -15,9 +15,15 @@ export interface ExecutorSettings {
 	tsPath: string;
 	tsArgs: string;
 	tsInject: string;
+	leanPath: string;
+	leanArgs: string;
+	leanInject: string;
 	luaPath: string;
 	luaArgs: string;
 	luaInject: string;
+	dartPath: string;
+	dartArgs: string;
+	dartInject: string;
 	csPath: string;
 	csArgs: string;
 	csInject: string;
@@ -29,6 +35,10 @@ export interface ExecutorSettings {
 	shellArgs: string;
 	shellFileExtension: string;
 	shellInject: string;
+	batchPath: string;
+	batchArgs: string;
+	batchFileExtension: string;
+	batchInject: string;
 	groovyPath: string;
 	groovyArgs: string;
 	groovyFileExtension: string;
@@ -79,18 +89,32 @@ export interface ExecutorSettings {
 	scalaArgs: string;
 	scalaFileExtension: string;
 	scalaInject: string;
+	racketPath: string;
+	racketArgs: string;
+	racketFileExtension: string;
+	racketInject: string;
+	fsharpPath: string;
+	fsharpArgs: string;
+	fsharpInject: "";
+	fsharpFileExtension: string;
 	cArgs: string;
 	cUseMain: boolean;
 	cInject: string;
+	rubyPath: string;
+	rubyArgs: string;
+	rubyInject: string;
 
 	jsInteractive: boolean;
 	tsInteractive: boolean;
 	csInteractive: boolean;
+	leanInteractive: boolean;
 	luaInteractive: boolean;
+	dartInteractive: boolean;
 	pythonInteractive: boolean;
 	cppInteractive: boolean;
 	prologInteractive: boolean;
 	shellInteractive: boolean;
+	batchInteractive: boolean;
 	bashInteractive: boolean;
 	groovyInteractive: boolean;
 	rInteractive: boolean;
@@ -102,7 +126,10 @@ export interface ExecutorSettings {
 	mathematicaInteractive: boolean;
 	haskellInteractive: boolean;
 	scalaInteractive: boolean;
+	racketInteractive: boolean;
+	fsharpInteractive: boolean;
 	cInteractive: boolean;
+	rubyInteractive: boolean;
 }
 
 
@@ -121,9 +148,15 @@ export const DEFAULT_SETTINGS: ExecutorSettings = {
 	tsPath: "ts-node",
 	tsArgs: "",
 	tsInject: "",
+	leanPath: "lean",
+	leanArgs: "",
+	leanInject: "",
 	luaPath: "lua",
 	luaArgs: "",
 	luaInject: "",
+	dartPath: "dart",
+	dartArgs: "",
+	dartInject: "",
 	csPath: "dotnet-script",
 	csArgs: "",
 	csInject: "",
@@ -135,6 +168,10 @@ export const DEFAULT_SETTINGS: ExecutorSettings = {
 	shellArgs: "",
 	shellFileExtension: "sh",
 	shellInject: "",
+	batchPath: "call",
+	batchArgs: "",
+	batchFileExtension: "bat",
+	batchInject: "",
 	groovyPath: "groovy",
 	groovyArgs: "",
 	groovyFileExtension: "groovy",
@@ -185,18 +222,32 @@ export const DEFAULT_SETTINGS: ExecutorSettings = {
 	scalaArgs: "",
 	scalaFileExtension: "scala",
 	scalaInject: "",
+	racketPath: "racket",
+	racketArgs: "",
+	racketFileExtension: "rkt",
+	racketInject: "#lang racket",
+	fsharpPath: "dotnet",
+	fsharpArgs: "",
+	fsharpInject: "",
+	fsharpFileExtension: "fsx",
 	cArgs: "",
 	cUseMain: true,
 	cInject: "",
+	rubyPath: "ruby",
+	rubyArgs: "",
+	rubyInject: "",
 
 	jsInteractive: true,
 	tsInteractive: false,
 	csInteractive: false,
+	leanInteractive: false,
 	luaInteractive: false,
+	dartInteractive: false,
 	pythonInteractive: true,
 	cppInteractive: false,
 	prologInteractive: false,
 	shellInteractive: false,
+	batchInteractive: false,
 	bashInteractive: false,
 	groovyInteractive: false,
 	rInteractive: false,
@@ -208,5 +259,8 @@ export const DEFAULT_SETTINGS: ExecutorSettings = {
 	mathematicaInteractive: false,
 	haskellInteractive: false,
 	scalaInteractive: false,
-	cInteractive: false
+	fsharpInteractive: false,
+	cInteractive: false,
+	racketInteractive: false,
+	rubyInteractive: false
 }