+++ src/transforms/Magic.ts
@@ -8,6 +8,7 @@
  * - `@vault`: Inserts the vault path as string.
  * - `@note`: Inserts the note path as string.
  * - `@title`: Inserts the note title as string.
+ * - `@theme`: Inserts the color theme; either `"light"` or `"dark"`. For use with images, inline plots, and `@html()`.
  */
 
 import * as os from "os";
@@ -17,8 +18,13 @@ import { TOGGLE_HTML_SIGIL } from "src/Outputter";
 const SHOW_REGEX = /@show\(["'](?<path>[^<>?*=!\n#()\[\]{}]+)["'](,\s*(?<width>\d+[\w%]+),?\s*(?<height>\d+[\w%]+))?(,\s*(?<align>left|center|right))?\)/g;
 const HTML_REGEX = /@html\((?<html>[^)]+)\)/g;
 const VAULT_REGEX = /@vault/g
+const VAULT_PATH_REGEX = /@vault_path/g
+const VAULT_URL_REGEX = /@vault_url/g
 const CURRENT_NOTE_REGEX = /@note/g;
+const CURRENT_NOTE_PATH_REGEX = /@note_path/g;
+const CURRENT_NOTE_URL_REGEX = /@note_url/g;
 const NOTE_TITLE_REGEX = /@title/g;
+const COLOR_THEME_REGEX = /@theme/g;
 
 // Regex that are only used by one language.
 const PYTHON_PLOT_REGEX = /^(plt|matplotlib.pyplot|pyplot)\.show\(\)/gm;
@@ -32,7 +38,11 @@ const R_PLOT_REGEX = /^plot\(.*\)/gm;
  * @returns The transformed source code.
  */
 export function insertVaultPath(source: string, vaultPath: string): string {
-	return source.replace(VAULT_REGEX, `"app://local/${vaultPath.replace(/\\/g, "/")}"`);
+	source = source.replace(VAULT_REGEX, `"app://local/${vaultPath.replace(/\\/g, "/")}"`);
+	source = source.replace(VAULT_URL_REGEX, `"app://local/${vaultPath.replace(/\\/g, "/")}"`);
+	source = source.replace(VAULT_PATH_REGEX, `"${vaultPath.replace(/\\/g, "/")}"`);
+
+	return source;
 }
 
 
@@ -44,7 +54,11 @@ export function insertVaultPath(source: string, vaultPath: string): string {
  * @returns The transformed source code.
  */
 export function insertNotePath(source: string, notePath: string): string {
-	return source.replace(CURRENT_NOTE_REGEX, `"app://local/${notePath.replace(/\\/g, "/")}"`);
+	source = source.replace(CURRENT_NOTE_REGEX, `"app://local/${notePath.replace(/\\/g, "/")}"`);
+	source = source.replace(CURRENT_NOTE_URL_REGEX, `"app://local/${notePath.replace(/\\/g, "/")}"`);
+	source = source.replace(CURRENT_NOTE_PATH_REGEX, `"${notePath.replace(/\\/g, "/")}"`);
+
+	return source;
 }
 
 
@@ -63,6 +77,16 @@ export function insertNoteTitle(source: string, noteTitle: string): string {
 	return source.replace(NOTE_TITLE_REGEX, `"${t}"`);
 }
 
+/**
+ * Parses the source code for the @theme command and replaces it with the colour theme.
+ *
+ * @param source The source code to parse.
+ * @param noteTitle The current colour theme.
+ * @returns The transformed source code.
+ */
+export function insertColorTheme(source: string, theme: string): string {
+	return source.replace(COLOR_THEME_REGEX, `"${theme}"`);
+}
 
 /**
  * Add the @show command to python. @show is only supported in python and javascript.